Enhancement: refactor gs_branches_diff_commit_history to gs_compare_commit to compare any commits
