+++ Default.sublime-keymap
@@ -930,15 +930,15 @@
         ]
     },
 
-    //////////////////////////////
-    // BRANCH COMMIT COMPARISON //
-    //////////////////////////////
+    ///////////////////////
+    // COMMIT COMPARISON //
+    ///////////////////////
 
     {
         "keys": ["enter"],
-        "command": "gs_branches_diff_commit_history_action",
+        "command": "gs_compare_commit_action",
         "context": [
-            { "key": "setting.git_savvy.branch_commit_history_view", "operator": "equal", "operand": true }
+            { "key": "setting.git_savvy.compare_commit_view", "operator": "equal", "operand": true }
         ]
     },
     {
@@ -946,7 +946,7 @@
         "command": "gs_log_graph_next_commit",
         "args": {"forward": true},
         "context": [
-            { "key": "setting.git_savvy.branch_commit_history_view", "operator": "equal", "operand": true }
+            { "key": "setting.git_savvy.compare_commit_view", "operator": "equal", "operand": true }
         ]
     },
     {
@@ -954,21 +954,21 @@
         "command": "gs_log_graph_next_commit",
         "args": {"forward": false},
         "context": [
-            { "key": "setting.git_savvy.branch_commit_history_view", "operator": "equal", "operand": true }
+            { "key": "setting.git_savvy.compare_commit_view", "operator": "equal", "operand": true }
         ]
     },
         {
         "keys": ["m"],
         "command": "gs_log_graph_toggle_more_info",
         "context": [
-            { "key": "setting.git_savvy.branch_commit_history_view", "operator": "equal", "operand": true }
+            { "key": "setting.git_savvy.compare_commit_view", "operator": "equal", "operand": true }
         ]
     },
     {
         "keys": ["r"],
-        "command": "gs_branches_diff_commit_history_refresh",
+        "command": "gs_ompare_commit_refresh",
         "context": [
-            { "key": "setting.git_savvy.branch_commit_history_view", "operator": "equal", "operand": true }
+            { "key": "setting.git_savvy.compare_commit_view", "operator": "equal", "operand": true }
         ]
     },
